+++ src/main/java/ru/skypro/lessons/springboot/springbootmvc/exceptions/EmployeeExceptionHandler.java
@@ -2,11 +2,11 @@ package ru.skypro.lessons.springboot.springbootmvc.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import java.io.IOException;
-import java.nio.file.AccessDeniedException;
 import java.sql.SQLException;
 
 @RestControllerAdvice
